YTG-#25 [상세페이지][모델] 모델 데이터 클래스 추가 및 수정
