+++ model/src/main/java/com/az/model/posts/AuthorData.kt
@@ -1,0 +1,8 @@
+package com.az.model.posts
+
+data class AuthorData(
+    val id: Int,
+    val identification: String,
+    val nickname: String,
+    val rating: String
+)
