+++ model/src/main/java/com/az/model/posts/SimplePageData.kt
@@ -1,0 +1,7 @@
+package com.az.model.posts
+
+data class SimplePageData(
+    val currentPage: Int,
+    val totalElements: Int,
+    val totalPages: Int
+)
