+++ model/src/main/java/com/az/model/posts/detail/comments/AddEditCommentData.kt
@@ -1,0 +1,5 @@
+package com.az.model.posts.detail.comments
+
+data class AddEditCommentData(
+    val comment: String
+)
