+++ model/src/main/java/com/az/model/posts/detail/comments/CommentData.kt
@@ -1,0 +1,10 @@
+package com.az.model.posts.detail.comments
+
+data class CommentData(
+    val content: String,
+    val createdDate: String,
+    val id: Int,
+    val modifiedDate: String,
+    val postId: Int,
+    val writer: String
+)
