+++ model/src/main/java/com/az/model/posts/detail/comments/CommentsData.kt
@@ -1,2 +1,8 @@
 package com.az.model.posts.detail.comments
 
+import com.az.model.posts.SimplePageData
+
+data class CommentsData(
+    val commentList: List<CommentData>,
+    val simplePage: SimplePageData
+)
